requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
et_xmlfile==2.0.0
fastapi==0.114.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
iniconfig==2.0.0
jose==1.0.0
motor==3.5.1
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pydantic==2.9.1
pydantic-settings==2.4.0
pydantic_core==2.23.3
pymongo==4.8.0
pymssql==2.3.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
